update make APT work
One thing I noticed is the "signal" we pass to matlab is a list type in Python. And matlab interprets it as a cell type variable and this causes some error.  The way I deal with it is to pass the path of signal and let matlab load it, since this variable remains same for every iteration